Added the support for the carousel
Added the support and the structures for the carousel. Still need to add the way to access without the use of the console. Also need to make it so that videos automatically stop when switching slides and have the image and video tags be centered on the screen instead of on the left side because of the ul